Less excpetions for set slot. Check for out-of-bounds before trying to set slots. Don't broadcast effect operations. Need to fix invisible, not working (not used).
