+++ src/MiNET/MiNET/Effects/Absorption.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class Absorption : Effect
+	{
+		public Absorption() : base(EffectType.Absorption)
+		{
+		}
+	}
+}
