+++ src/MiNET/MiNET/Effects/Blindness.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Blindness : Effect
+	{
+		public Blindness() : base(EffectType.Blindness)
+		{
+		}
+	}
+}
