+++ src/MiNET/MiNET/Effects/FireResistance.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class FireResistance : Effect
+	{
+		public FireResistance() : base(EffectType.FireResistance)
+		{
+		}
+	}
+}
