+++ src/MiNET/MiNET/Effects/Glowing.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Glowing : Effect
+	{
+		public Glowing() : base(EffectType.Glowing)
+		{
+		}
+	}
+}
