+++ src/MiNET/MiNET/Effects/Haste.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Haste : Effect
+	{
+		public Haste() : base(EffectType.Haste)
+		{
+		}
+	}
+}
