+++ src/MiNET/MiNET/Effects/HealthBoost.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class HealthBoost : Effect
+	{
+		public HealthBoost() : base(EffectType.HealthBoost)
+		{
+		}
+	}
+}
