+++ src/MiNET/MiNET/Effects/Hunger.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class Hunger : Effect
+	{
+		public Hunger() : base(EffectType.Hunger)
+		{
+		}
+	}
+}
