+++ src/MiNET/MiNET/Effects/InstandHealth.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class InstandHealth : Effect
+	{
+		public InstandHealth() : base(EffectType.InstandHealth)
+		{
+		}
+	}
+}
