+++ src/MiNET/MiNET/Effects/InstantDamage.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class InstantDamage : Effect
+	{
+		public InstantDamage() : base(EffectType.InstantDamage)
+		{
+		}
+	}
+}
