+++ src/MiNET/MiNET/Effects/Invisibility.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Invisibility : Effect
+	{
+		public Invisibility() : base(EffectType.Invisibility)
+		{
+		}
+	}
+}
