+++ src/MiNET/MiNET/Effects/JumpBoost.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class JumpBoost : Effect
+	{
+		public JumpBoost() : base(EffectType.JumpBoost)
+		{
+		}
+	}
+}
