+++ src/MiNET/MiNET/Effects/Levitation.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Levitation : Effect
+	{
+		public Levitation() : base(EffectType.Levitation)
+		{
+		}
+	}
+}
