+++ src/MiNET/MiNET/Effects/MiningFatigue.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class MiningFatigue : Effect
+	{
+		public MiningFatigue() : base(EffectType.MiningFatigue)
+		{
+		}
+	}
+}
