+++ src/MiNET/MiNET/Effects/Nausea.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Nausea : Effect
+	{
+		public Nausea() : base(EffectType.Nausea)
+		{
+		}
+	}
+}
