+++ src/MiNET/MiNET/Effects/NightVision.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class NightVision : Effect
+	{
+		public NightVision() : base(EffectType.NightVision)
+		{
+		}
+	}
+}
