+++ src/MiNET/MiNET/Effects/Poison.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Poison : Effect
+	{
+		public Poison() : base(EffectType.Poison)
+		{
+		}
+	}
+}
