+++ src/MiNET/MiNET/Effects/Regeneration.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Regeneration : Effect
+	{
+		public Regeneration() : base(EffectType.Regeneration)
+		{
+		}
+	}
+}
