+++ src/MiNET/MiNET/Effects/Resistance.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Resistance : Effect
+	{
+		public Resistance() : base(EffectType.Resistance)
+		{
+		}
+	}
+}
