+++ src/MiNET/MiNET/Effects/Saturation.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class Saturation : Effect
+	{
+		public Saturation() : base(EffectType.Saturation)
+		{
+		}
+	}
+}
