+++ src/MiNET/MiNET/Effects/Slowness.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class Slowness : Effect
+	{
+		public Slowness() : base(EffectType.Slowness)
+		{
+		}
+	}
+}
