+++ src/MiNET/MiNET/Effects/Speed.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Speed : Effect
+	{
+		public Speed() : base(EffectType.Speed)
+		{
+		}
+	}
+}
