+++ src/MiNET/MiNET/Effects/Strength.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class Strength : Effect
+	{
+		public Strength() : base(EffectType.Strenght)
+		{
+		}
+	}
+}
