+++ src/MiNET/MiNET/Effects/WaterBreathing.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class WaterBreathing : Effect
+	{
+		public WaterBreathing() : base(EffectType.WaterBreathing)
+		{
+		}
+	}
+}
