+++ src/MiNET/MiNET/Effects/Weakness.cs
@@ -1,0 +1,9 @@
+namespace MiNET.Effects
+{
+	public class Weakness : Effect
+	{
+		public Weakness() : base(EffectType.Weakness)
+		{
+		}
+	}
+}
