+++ src/MiNET/MiNET/Effects/Wither.cs
@@ -1,0 +1,9 @@
+﻿namespace MiNET.Effects
+{
+	public class Wither : Effect
+	{
+		public Wither() : base(EffectType.Wither)
+		{
+		}
+	}
+}
