+++ src/MiNET/MiNET/MiNET.csproj
@@ -127,7 +127,32 @@
     <Compile Include="Blocks\Tnt.cs" />
     <Compile Include="ConnectionState.cs" />
     <Compile Include="Crafting\Recipe.cs" />
+    <Compile Include="Effects\Absorption.cs" />
+    <Compile Include="Effects\Blindness.cs" />
     <Compile Include="Effects\Effect.cs" />
+    <Compile Include="Effects\FireResistance.cs" />
+    <Compile Include="Effects\Glowing.cs" />
+    <Compile Include="Effects\Haste.cs" />
+    <Compile Include="Effects\HealthBoost.cs" />
+    <Compile Include="Effects\Hunger.cs" />
+    <Compile Include="Effects\InstandHealth.cs" />
+    <Compile Include="Effects\InstantDamage.cs" />
+    <Compile Include="Effects\Invisibility.cs" />
+    <Compile Include="Effects\JumpBoost.cs" />
+    <Compile Include="Effects\Levitation.cs" />
+    <Compile Include="Effects\MiningFatigue.cs" />
+    <Compile Include="Effects\Nausea.cs" />
+    <Compile Include="Effects\NightVision.cs" />
+    <Compile Include="Effects\Poison.cs" />
+    <Compile Include="Effects\Regeneration.cs" />
+    <Compile Include="Effects\Resistance.cs" />
+    <Compile Include="Effects\Saturation.cs" />
+    <Compile Include="Effects\Slowness.cs" />
+    <Compile Include="Effects\Speed.cs" />
+    <Compile Include="Effects\Strength.cs" />
+    <Compile Include="Effects\WaterBreathing.cs" />
+    <Compile Include="Effects\Weakness.cs" />
+    <Compile Include="Effects\Wither.cs" />
     <Compile Include="Entities\Arrow.cs" />
     <Compile Include="Entities\Egg.cs" />
     <Compile Include="Entities\Entity.cs" />
